requirements.txt
asgiref==3.7.2
black==23.3.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-completion==0.5.2
colorama==0.4.6
coverage==7.2.5
crayons==0.4.0
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==1.3.0
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
exceptiongroup==1.2.0
factory-boy==3.2.1
Faker==23.2.1
filelock==3.13.1
gunicorn==20.1.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.3.2
pycparser==2.21
pycycle==0.0.8
PyJWT==2.8.0
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.0.267
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
whitenoise==6.4.0
asgiref==3.7.2
black==23.3.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-completion==0.5.2
colorama==0.4.6
coverage==7.2.5
crayons==0.4.0
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==1.3.0
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
exceptiongroup==1.2.0
factory-boy==3.2.1
Faker==23.2.1
filelock==3.13.1
gunicorn==20.1.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.3.2
pycparser==2.21
pycycle==0.0.8
PyJWT==2.8.0
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.0.267
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
whitenoise==6.4.0
